Modularize login, identify endpoints, ignore sessions in .gitignore
