+++ .gitignore
@@ -34,3 +34,6 @@ node_modules
 
 # parcel-bundler cache (https://parceljs.org/)
 .cache
+
+# ignore backend sessions (and Jeff Sessions?)
+sessions
