+++ modules/api/server.js
@@ -10,247 +10,10 @@ ABOUT THIS FILE
 
 This file is in the process of being deprecated and modularized.
 For those working on this process, please use register.js as a guide
-for how to take one of these sections and modularize it into it's own
-file. This will make things more easy to manage long-term.
-
-Please refrain from deleting or changing server.js until the process
-is complete. webSericesBackend.js should be the new entry point for
-the server (as opposed to this file) once this is done.
-
-Also, FundsTransfer.js and FixBrokenPayments.js are to be initialized
-and managed as part of the server.
+for how to take one of these sections and modularize it into it's 
+own file. This will make things more easy to manage long-term.
 
 */
-
-const express = require('express')
-const session = require('express-session')
-const FileStore = require('session-file-store')(session)
-const bodyParser = require('body-parser')
-const url = require('url')
-const sha256 = require('sha256')
-const mysql = require('mysql')
-const BITBOXCli = require('bitbox-cli/lib/bitbox-cli').default
-const bch = new BITBOXCli()
-
-// define the app
-const app = express()
-
-// se up bodyParser tor form data handling
-app.use(bodyParser.json())
-app.use(bodyParser.urlencoded({ extended: true }))
-
-// set up persistant session
-app.use(
-  session({
-    name: 'gateway-session',
-    // TODO: use environment variable
-    secret: 'sajdfpaioewypjmpasodjw,adfjwoiejfo',
-    saveUninitialized: true,
-    resave: true,
-    store: new FileStore(),
-  }),
-)
-
-// enable CORS to allow for API calls from other sites
-app.use((req, res, next) => {
-  res.header('Access-Control-Allow-Origin', '*')
-  res.header(
-    'Access-Control-Allow-Headers',
-    'Origin, X-Requested-With, Content-Type, Accept',
-  )
-  next()
-})
-
-// listen on port 8080
-app.listen(8080, () => {
-  console.log('Gateway listening on port 8080')
-})
-
-// connect to the database
-var conn = mysql.createConnection({
-  host: 'localhost',
-  user: 'gateway',
-  // TODO: use environment variales
-  password: 'gateway123',
-  database: 'gateway',
-})
-conn.connect((err) => {
-  if (err) {
-    throw err
-  }
-})
-
-app.post('/api/register', (req, res) => {
-  if (!req.session.address || req.session.address.length < 20) {
-    res.send('Make sure to give an address before trying to register')
-  } else if (!bch.Address.isCashAddress(req.session.address)) {
-    res.send('Make sure the provided addres is a valid CashAddress!')
-  } else if (!req.body.password) {
-    res.send('Please provide a password')
-  } else if (req.body.password.length < 12) {
-    res.send('Password must be at least 12 characters')
-  } else if (req.body.password.toString().indexOf(' ') !== -1) {
-    res.send('No spaces are allowed in your password')
-  } else {
-    // initial checks passed, make sure address is not in the database already
-    var sql = 'select payoutAddress from users where payoutAddress = ?'
-    conn.query(sql, [req.session.address], (err, result) => {
-      if (err) {
-        throw err
-      }
-      if (result.length !== 0) {
-        res.send('This address has already been registered')
-      } else {
-        // no duplicate found, create a new row in the database
-        var merchantID = sha256(req.session.address).substr(0, 16)
-        var passwordSalt = sha256(require('crypto').randomBytes(32))
-        var passwordHash = sha256(req.body.password + passwordSalt)
-        var sql = `insert into users
-        (payoutAddress, merchantID, password, salt)
-        values
-        (?, ?, ?, ?)`
-        conn.query(
-          sql,
-          [req.session.address, merchantID, passwordHash, passwordSalt],
-          (err, result) => {
-            if (err) {
-              throw err
-            }
-            // registration complete, set up session and print OK message
-            req.session.merchantID = merchantID
-            req.session.username = ''
-            req.session.loggedIn = true
-            res.send('ok')
-          },
-        )
-      }
-    })
-  }
-})
-
-app.get('/api/identify', (req, res) => {
-  // parse the provided data
-  const query = url.parse(req.url, true).query
-  if (query.type !== 'address' && query.type !== 'username') {
-    res.send('Error: invalid type (must be address or username)')
-  } else if (query.type === 'address') {
-    // log in by addresss
-    // a variable to hold the converted and valid addresss
-    var address = false
-    // make sure provided address is valid
-    try {
-      if (
-        !bch.Address.isCashAddress(query.value) &&
-        !bch.Address.isLegacyAddress(query.value)
-      ) {
-        // invalid address
-        res.send('Error: Invalid address (please use CashAddr or Legacy)')
-      } else if (bch.Address.isCashAddress(query.value)) {
-        // cash address
-        address = query.value
-        // check if prefix was given
-        if (!address.startsWith('bitcoincash:')) {
-          address = 'bitcoincash:' + address
-        }
-      } else {
-        // legacy address
-        address = bch.Address.toCashAddress(query.value, true)
-      }
-    } catch (e) {
-      res.send('Invalid address (use CashAddr or Legacy)')
-    }
-    if (address !== false) {
-      // search the database for the given address
-      var sql = 'select payoutAddress from users where payoutAddress = ?'
-      conn.query(sql, [address], (err, result) => {
-        if (err) {
-          throw err
-        }
-        // save the address to persistant session
-        req.session.address = address
-        if (result.length !== 1) {
-          res.send('register')
-        } else {
-          res.send('login')
-        }
-      })
-    }
-  } else {
-    // log in by username
-    if (query.value.length < 10) {
-      res.send('No match')
-    } else {
-      var sql = 'select username from users where username = ?'
-      conn.query(sql, [query.value], (err, result) => {
-        if (result.length !== 1) {
-          res.send('No match')
-        } else {
-          req.session.username = query.value
-          res.send('login')
-        }
-      })
-    }
-  }
-})
-
-app.get('/api/login', (req, res) => {
-  // parse the provided data
-  const query = url.parse(req.url, true).query
-  if (!req.session.address && !req.session.username) {
-    res.send('Provide an address or username first')
-  } else if (query.password.length < 12) {
-    res.send('Incorrect password')
-  } else if (req.session.address) {
-    // prefer logging in by address
-    var sql = 'select * from users where payoutAddress = ?'
-    conn.query(sql, [req.session.address], (err, result) => {
-      if (err) {
-        throw err
-      }
-      if (result.length !== 1) {
-        res.send('Incorrect address')
-      } else {
-        var user = result[0]
-        var passwordHash = sha256(query.password + user.salt)
-        if (user.password === passwordHash) {
-          // login successful, set up session variables
-          req.session.address = user.payoutAddress
-          req.session.username = user.username
-          req.session.merchantID = user.merchantID
-          req.session.loggedIn = true
-          res.send('ok')
-        } else {
-          res.send('Incorrect password')
-        }
-      }
-    })
-  } else {
-    // search by username
-    console.log(req.session)
-    var sql = 'select * from users where username = ?'
-    conn.query(sql, [req.session.username], (err, result) => {
-      if (err) {
-        throw err
-      }
-      if (result.length !== 1) {
-        res.send('Incorrect username')
-      } else {
-        var user = result[0]
-        var passwordHash = sha256(query.password + user.salt)
-        if (user.password === passwordHash) {
-          // login successful, set up session variables
-          req.session.address = user.payoutAddress
-          req.session.username = user.username
-          req.session.merchantID = user.merchantID
-          req.session.loggedIn = true
-          res.send('ok')
-        } else {
-          res.send('Incorrect password')
-        }
-      }
-    })
-  }
-})
 
 app.get('/api/pay', (req, res) => {
   const query = url.parse(req.url, true).query
